refactor(discord): extract stream loop tool arms
