+++ ARCHITECTURE.md
@@ -701,6 +701,8 @@ src/
 │   │   │   │   └── completion_postgres.rs
 │   │   │   ├── runtime_handoff_loop/
 │   │   │   │   └── guarded_save.rs
+│   │   │   ├── stream_loop/
+│   │   │   │   └── tool_arms.rs
 │   │   │   ├── tmux_runtime/
 │   │   │   │   ├── interrupt_policy.rs
 │   │   │   │   ├── pid_exit.rs
