+++ scripts/audit_maintainability_config.toml
@@ -32,7 +32,8 @@
 "src/services/discord/tmux_watcher/no_result_exits.rs" = 814 # 2026-07-08 #4229 S6: approved contiguous verbatim extraction of the no-result exit handler (fresh ready-for-input idle structural finalize / tmux-death break / cancel-shutdown break / ready-for-input stall notice+dispatch-fail / #3419 watchdog-timeout finalize); below giant threshold (<1000), but irreducible as one review-verified move.
 "src/services/discord/tmux_watcher/**" = 700
 "src/services/discord/turn_bridge/headless_delivery.rs" = 700
-"src/services/discord/turn_bridge/stream_loop.rs" = 1638 # 2026-07-08 #4230 S6: registered giant-file, plan-marked irreducible stream loop shell (event arms share placeholder/background/transcript/inflight state); tracked in giant_file_registry.toml.
+"src/services/discord/turn_bridge/stream_loop.rs" = 1291 # 2026-07-10 #4230 S7: lowered after approved contiguous verbatim extraction of ToolUse/ToolResult/TaskNotification arms to stream_loop/tool_arms.rs; still tracked in giant_file_registry.toml.
+"src/services/discord/turn_bridge/stream_loop/tool_arms.rs" = 684 # 2026-07-10 #4230 S7 approved contiguous verbatim extraction of ToolUse/ToolResult/TaskNotification stream-loop arms.
 "src/services/discord/turn_bridge/status_panel.rs" = 700
 "src/services/discord/turn_bridge/voice_completion.rs" = 700
